task6

+++ 0x00-hello_world/6-size.c
@@ -8,11 +8,11 @@
 int main(void)
 {
 
-  printf("The size of a char is: %lu byte(s)\n", (unsigned long)sizeof(char));
-  printf("The size of an int is: %lu byte(s)\n", (unsigned long)sizeof(int));
-  printf("The size of a long int is: %lu byte(s)\n", (unsigned long)sizeof(long int));
-  printf("The size of a long long int is: %lu bytes(s)\n", (unsigned long)sizeof(long long int));
-  printf("The size of a float is: %lu byte(s)\n", (unsigned long)sizeof(float));
+  printf("The size of a char is: %lu byte(s)\n", sizeof(char));
+  printf("The size of an int is: %lu byte(s)\n", sizeof(int));
+  printf("The size of a long int is: %lu byte(s)\n", sizeof(long int));
+  printf("The size of a long long int is: %lu bytes(s)\n", sizeof(long long int));
+  printf("The size of a float is: %lu byte(s)\n", sizeof(float));
 
   return (0);
 }
